+++ packages/cli/src/handlers/download.ts
@@ -4,6 +4,7 @@ import {
     ApiChartAsCodeListResponse,
     ApiChartAsCodeUpsertResponse,
     ApiDashboardAsCodeListResponse,
+    assertUnreachable,
     AuthorizationError,
     ChartAsCode,
     DashboardAsCode,
@@ -13,6 +14,7 @@ import {
 import { promises as fs } from 'fs';
 import * as yaml from 'js-yaml';
 import * as path from 'path';
+import { LightdashAnalytics } from '../analytics/analytics';
 import { getConfig } from '../config';
 import GlobalState from '../globalState';
 import * as styles from '../styles';
@@ -52,7 +54,7 @@ const dumpIntoFiles = async (
 ) => {
     const outputDir = path.join(process.cwd(), DOWNLOAD_FOLDER, folder);
 
-    console.info(`Writing ${items.length} ${folder} into ${outputDir}`);
+    GlobalState.debug(`Writing ${items.length} ${folder} into ${outputDir}`);
     // Make directory
     const created = await fs.mkdir(outputDir, { recursive: true });
     if (created) console.info(`Created new folder: ${outputDir} `);
@@ -137,59 +139,156 @@ export const downloadHandler = async (
         );
     }
 
-    // If any filter is provided, we skip those items without filters
-    // eg: if a --charts filter is provided, we skip dashboards if no --dashboards filter is provided
-    const hasFilters =
-        options.charts.length > 0 || options.dashboards.length > 0;
+    // For analytics
+    let chartTotal: number | undefined;
+    let dashboardTotal: number | undefined;
+    const start = Date.now();
 
-    // Download charts
-    if (hasFilters && options.charts.length === 0) {
-        console.info(styles.warning(`No charts filters provided, skipping`));
-    } else {
-        GlobalState.debug(`Downloading charts`);
-        const chartFilters = parseContentFilters(options.charts);
+    await LightdashAnalytics.track({
+        event: 'download.started',
+        properties: {
+            userId: config.user?.userUuid,
+            organizationId: config.user?.organizationUuid,
+            projectId,
+        },
+    });
+    try {
+        // If any filter is provided, we skip those items without filters
+        // eg: if a --charts filter is provided, we skip dashboards if no --dashboards filter is provided
+        const hasFilters =
+            options.charts.length > 0 || options.dashboards.length > 0;
 
-        const chartsAsCode = await lightdashApi<
-            ApiChartAsCodeListResponse['results']
-        >({
-            method: 'GET',
-            url: `/api/v1/projects/${projectId}/charts/code${chartFilters}`,
-            body: undefined,
+        // Download charts
+        if (hasFilters && options.charts.length === 0) {
+            console.info(
+                styles.warning(`No charts filters provided, skipping`),
+            );
+        } else {
+            const spinner = GlobalState.startSpinner(`Downloading charts`);
+            const chartFilters = parseContentFilters(options.charts);
+            let chartsAsCode: ApiChartAsCodeListResponse['results'];
+            let offset = 0;
+            do {
+                GlobalState.debug(
+                    `Downloading charts with offset "${offset}" and filters "${chartFilters}"`,
+                );
+
+                const queryParams = chartFilters
+                    ? `${chartFilters}&offset=${offset}`
+                    : `?offset=${offset}`;
+                chartsAsCode = await lightdashApi<
+                    ApiChartAsCodeListResponse['results']
+                >({
+                    method: 'GET',
+                    url: `/api/v1/projects/${projectId}/charts/code${queryParams}`,
+                    body: undefined,
+                });
+                spinner.start(
+                    `Downloaded ${chartsAsCode.offset} of ${chartsAsCode.total} charts`,
+                );
+                chartsAsCode.missingIds.forEach((missingId) => {
+                    console.warn(
+                        styles.warning(`\nNo chart with id "${missingId}"`),
+                    );
+                });
+
+                await dumpIntoFiles('charts', chartsAsCode.charts);
+                offset = chartsAsCode.offset;
+            } while (chartsAsCode.offset < chartsAsCode.total);
+
+            chartTotal = chartsAsCode.total;
+            spinner.succeed(`Downloaded ${chartsAsCode.total} charts`);
+        }
+
+        // Download dashboards
+        if (hasFilters && options.dashboards.length === 0) {
+            console.info(
+                styles.warning(`No dashboards filters provided, skipping`),
+            );
+        } else {
+            const spinner = GlobalState.startSpinner(`Downloading dashboards`);
+
+            const dashboardFilters = parseContentFilters(options.dashboards);
+            let offset = 0;
+
+            let dashboardsAsCode: ApiDashboardAsCodeListResponse['results'];
+            do {
+                GlobalState.debug(
+                    `Downloading dashboards with offset "${offset}" and filters "${dashboardFilters}"`,
+                );
+
+                const queryParams = dashboardFilters
+                    ? `${dashboardFilters}&offset=${offset}`
+                    : `?offset=${offset}`;
+                dashboardsAsCode = await lightdashApi<
+                    ApiDashboardAsCodeListResponse['results']
+                >({
+                    method: 'GET',
+                    url: `/api/v1/projects/${projectId}/dashboards/code${queryParams}`,
+                    body: undefined,
+                });
+
+                dashboardsAsCode.missingIds.forEach((missingId) => {
+                    console.warn(
+                        styles.warning(`\nNo dashboard with id "${missingId}"`),
+                    );
+                });
+                spinner?.start(
+                    `Downloaded ${dashboardsAsCode.offset} of ${dashboardsAsCode.total} dashboards`,
+                );
+
+                await dumpIntoFiles('dashboards', dashboardsAsCode.dashboards);
+                offset = dashboardsAsCode.offset;
+            } while (dashboardsAsCode.offset < dashboardsAsCode.total);
+
+            dashboardTotal = dashboardsAsCode.total;
+            spinner.succeed(`Downloaded ${dashboardsAsCode.total} dashboards`);
+        }
+
+        const end = Date.now();
+
+        await LightdashAnalytics.track({
+            event: 'download.completed',
+            properties: {
+                userId: config.user?.userUuid,
+                organizationId: config.user?.organizationUuid,
+                projectId,
+                chartsNum: chartTotal,
+                dashboardsNum: dashboardTotal,
+                timeToCompleted: (end - start) / 1000, // in seconds
+            },
         });
-
-        chartsAsCode.missingIds.forEach((missingId) => {
-            console.warn(styles.warning(`No chart with id "${missingId}"`));
+    } catch (error) {
+        console.error(styles.error(`\nError downloading ${error}`));
+        await LightdashAnalytics.track({
+            event: 'download.error',
+            properties: {
+                userId: config.user?.userUuid,
+                organizationId: config.user?.organizationUuid,
+                projectId,
+                error: `${error}`,
+            },
         });
-
-        await dumpIntoFiles('charts', chartsAsCode.charts);
-    }
-    // Download dashboards
-    if (hasFilters && options.dashboards.length === 0) {
-        console.info(
-            styles.warning(`No dashboards filters provided, skipping`),
-        );
-    } else {
-        GlobalState.debug(`Downloading dashboards`);
-        const dashboardFilters = parseContentFilters(options.dashboards);
-
-        const dashboardsAsCode = await lightdashApi<
-            ApiDashboardAsCodeListResponse['results']
-        >({
-            method: 'GET',
-            url: `/api/v1/projects/${projectId}/dashboards/code${dashboardFilters}`,
-            body: undefined,
-        });
-
-        dashboardsAsCode.missingIds.forEach((missingId) => {
-            console.warn(styles.warning(`No dashboard with id "${missingId}"`));
-        });
-
-        await dumpIntoFiles('dashboards', dashboardsAsCode.dashboards);
     }
 
     // TODO delete files if chart don't exist ?*/
 };
 
+const getPromoteAction = (action: PromotionAction) => {
+    switch (action) {
+        case PromotionAction.CREATE:
+            return 'created';
+        case PromotionAction.UPDATE:
+            return 'updated';
+        case PromotionAction.DELETE:
+            return 'deleted';
+        case PromotionAction.NO_CHANGES:
+            return 'skipped';
+        default:
+            assertUnreachable(action, `Unknown promotion action: ${action}`);
+    }
+    return 'skipped';
+};
 const storeUploadChanges = (
     changes: Record<string, number>,
     promoteChanges: PromotionChanges,
@@ -201,10 +300,7 @@ const storeUploadChanges = (
             promoteChanges[resource];
         return promotions.reduce<Record<string, number>>(
             (acc, promoteChange) => {
-                const action =
-                    promoteChange.action === PromotionAction.NO_CHANGES
-                        ? 'skipped'
-                        : promoteChange.action;
+                const action = getPromoteAction(promoteChange.action);
                 const key = `${resource} ${action}`;
                 acc[key] = (acc[key] ?? 0) + 1;
                 return acc;
@@ -244,7 +340,9 @@ const upsertResources = async <T extends ChartAsCode | DashboardAsCode>(
     changes: Record<string, number>,
     force: boolean,
     slugs: string[],
-) => {
+): Promise<{ changes: Record<string, number>; total: number }> => {
+    const config = await getConfig();
+
     const items = await readCodeFiles<T>(type);
 
     console.info(`Found ${items.length} ${type} files`);
@@ -303,10 +401,21 @@ const upsertResources = async <T extends ChartAsCode | DashboardAsCode>(
         } catch (error) {
             changes[`${type} with errors`] =
                 (changes[`${type} with errors`] ?? 0) + 1;
-            console.error(`Error upserting ${type}`, error);
+            console.error(styles.error(`Error upserting ${type}: ${error}`));
+
+            await LightdashAnalytics.track({
+                event: 'download.error',
+                properties: {
+                    userId: config.user?.userUuid,
+                    organizationId: config.user?.organizationUuid,
+                    projectId,
+                    type,
+                    error: `${error}`,
+                },
+            });
         }
     }
-    return changes;
+    return { changes, total: filteredItems.length };
 };
 
 export const uploadHandler = async (
@@ -329,21 +438,84 @@ export const uploadHandler = async (
     }
 
     let changes: Record<string, number> = {};
+    // For analytics
+    let chartTotal: number | undefined;
+    let dashboardTotal: number | undefined;
+    const start = Date.now();
 
-    changes = await upsertResources<ChartAsCode>(
-        'charts',
-        projectId,
-        changes,
-        options.force,
-        options.charts,
-    );
-    changes = await upsertResources<DashboardAsCode>(
-        'dashboards',
-        projectId,
-        changes,
-        options.force,
-        options.dashboards,
-    );
+    await LightdashAnalytics.track({
+        event: 'upload.started',
+        properties: {
+            userId: config.user?.userUuid,
+            organizationId: config.user?.organizationUuid,
+            projectId,
+        },
+    });
 
-    logUploadChanges(changes);
+    try {
+        // If any filter is provided, we skip those items without filters
+        // eg: if a --charts filter is provided, we skip dashboards if no --dashboards filter is provided
+        const hasFilters =
+            options.charts.length > 0 || options.dashboards.length > 0;
+
+        if (hasFilters && options.charts.length === 0) {
+            console.info(
+                styles.warning(`No charts filters provided, skipping`),
+            );
+        } else {
+            const { changes: chartChanges, total } =
+                await upsertResources<ChartAsCode>(
+                    'charts',
+                    projectId,
+                    changes,
+                    options.force,
+                    options.charts,
+                );
+            changes = chartChanges;
+            chartTotal = total;
+        }
+
+        if (hasFilters && options.dashboards.length === 0) {
+            console.info(
+                styles.warning(`No dashboard filters provided, skipping`),
+            );
+        } else {
+            const { changes: dashboardChanges, total } =
+                await upsertResources<DashboardAsCode>(
+                    'dashboards',
+                    projectId,
+                    changes,
+                    options.force,
+                    options.dashboards,
+                );
+            changes = dashboardChanges;
+            dashboardTotal = total;
+        }
+        const end = Date.now();
+
+        await LightdashAnalytics.track({
+            event: 'upload.completed',
+            properties: {
+                userId: config.user?.userUuid,
+                organizationId: config.user?.organizationUuid,
+                projectId,
+                chartsNum: chartTotal,
+                dashboardsNum: dashboardTotal,
+                timeToCompleted: (end - start) / 1000, // in seconds
+            },
+        });
+
+        logUploadChanges(changes);
+    } catch (error) {
+        console.error(styles.error(`\nError downloading ${error}`));
+        await LightdashAnalytics.track({
+            event: 'download.error',
+            properties: {
+                userId: config.user?.userUuid,
+                organizationId: config.user?.organizationUuid,
+                projectId,
+                error: `${error}`,
+            },
+        });
+    }
 };