+++ packages/frontend/src/components/VisualizationConfigs/PieChartConfig/PieChartDisplayConfig.tsx
@@ -1,4 +1,5 @@
 import {
+    PieChartLegendLabelMaxLengthDefault,
     PieChartLegendPositions,
     type PieChartLegendPosition,
 } from '@lightdash/common';
@@ -8,6 +9,7 @@ import {
     SegmentedControl,
     Stack,
     Switch,
+    TextInput,
 } from '@mantine/core';
 import React from 'react';
 import { isPieVisualizationConfig } from '../../LightdashVisualization/VisualizationConfigPie';
@@ -24,6 +26,8 @@ export const Display: React.FC = () => {
         toggleShowLegend,
         legendPosition,
         legendPositionChange,
+        legendMaxItemLength,
+        legendMaxItemLengthChange,
     } = visualizationConfig.chartConfig;
 
     return (
@@ -36,22 +40,43 @@ export const Display: React.FC = () => {
             </Config>
 
             <Collapse in={showLegend}>
-                <Group spacing="xs">
-                    <Config.Label>Orientation</Config.Label>
-                    <SegmentedControl
-                        name="orient"
-                        value={legendPosition}
-                        onChange={(val: PieChartLegendPosition) =>
-                            legendPositionChange(val)
-                        }
-                        data={Object.entries(PieChartLegendPositions).map(
-                            ([position, label]) => ({
-                                label,
-                                value: position,
-                            }),
-                        )}
-                    />
-                </Group>
+                <Stack spacing="xs">
+                    <Group spacing="xs">
+                        <Config.Label>Orientation</Config.Label>
+                        <SegmentedControl
+                            name="orient"
+                            value={legendPosition}
+                            onChange={(val: PieChartLegendPosition) =>
+                                legendPositionChange(val)
+                            }
+                            data={Object.entries(PieChartLegendPositions).map(
+                                ([position, label]) => ({
+                                    label,
+                                    value: position,
+                                }),
+                            )}
+                        />
+                    </Group>
+                    <Group>
+                        <Config.Label>Label max length</Config.Label>
+                        <TextInput
+                            type="number"
+                            value={legendMaxItemLength}
+                            placeholder={PieChartLegendLabelMaxLengthDefault.toString()}
+                            onChange={(e) => {
+                                const parsedNumber = Number.parseInt(
+                                    e.target.value,
+                                    10,
+                                );
+                                legendMaxItemLengthChange(
+                                    !Number.isNaN(parsedNumber)
+                                        ? parsedNumber
+                                        : undefined,
+                                );
+                            }}
+                        />
+                    </Group>
+                </Stack>
             </Collapse>
         </Stack>
     );